--- aws-sdk-v2/src/testApache4/groovy/io/micronaut/aws/sdk/v2/client/apache/ApacheClientConfigurationSpec.groovy
@@ -1,19 +1,0 @@
-package io.micronaut.aws.sdk.v2.client.apache
-
-import io.micronaut.context.ApplicationContext
-import io.micronaut.context.annotation.BootstrapContextCompatible
-import spock.lang.Specification
-
-class ApacheClientConfigurationSpec extends Specification {
-
-    void 'ApacheClientConfiguration is annotated with BootstrapContextCompatible'() {
-        given:
-        ApplicationContext context = ApplicationContext.run()
-
-        expect:
-        context.getBeanDefinition(ApacheClientConfiguration).getAnnotationNameByStereotype(BootstrapContextCompatible).isPresent()
-
-        cleanup:
-        context.close()
-    }
-}
