+++ settings.gradle.kts
@@ -52,6 +52,10 @@ include("test-suite-aws-sdk-v2")
 include("test-suite-graal")
 include("test-suite-graal-logging")
 include("test-suite-groovy")
+include("test-suite-groovy-aws-sdk-v2-http-client-apache4")
+include("test-suite-groovy-aws-sdk-v2-http-client-apache5")
+include("test-suite-groovy-aws-sdk-v2-http-client-netty")
+include("test-suite-groovy-aws-sdk-v2-http-client-urlconnection")
 include("test-suite-function-client-aws-groovy")
 include("test-suite-http-server-tck-function-aws-api-gateway-proxy-alb")
 include("test-suite-http-server-tck-function-aws-api-gateway-proxy-payloadv1")
