+++ test-suite-groovy-aws-sdk-v2-http-client-apache4/src/test/groovy/io/micronaut/aws/sdk/v2/client/apache4/ApacheClientConfigurationSpec.groovy
@@ -1,0 +1,20 @@
+package io.micronaut.aws.sdk.v2.client.apache4
+
+import io.micronaut.aws.sdk.v2.client.apache.ApacheClientConfiguration
+import io.micronaut.context.ApplicationContext
+import io.micronaut.context.annotation.BootstrapContextCompatible
+import spock.lang.Specification
+
+class ApacheClientConfigurationSpec extends Specification {
+
+    void 'ApacheClientConfiguration is annotated with BootstrapContextCompatible'() {
+        given:
+        ApplicationContext context = ApplicationContext.run()
+
+        expect:
+        context.getBeanDefinition(ApacheClientConfiguration).getAnnotationNameByStereotype(BootstrapContextCompatible).isPresent()
+
+        cleanup:
+        context.close()
+    }
+}
