+++ test-suite-groovy-aws-sdk-v2-http-client-apache5/src/test/groovy/io/micronaut/aws/sdk/v2/client/apache5/Apache5ServiceClientSpec.groovy
@@ -1,0 +1,23 @@
+package io.micronaut.aws.sdk.v2.client.apache5
+
+import io.micronaut.context.BeanContext
+import io.micronaut.context.annotation.Property
+import io.micronaut.test.extensions.spock.annotation.MicronautTest
+import jakarta.inject.Inject
+import software.amazon.awssdk.services.s3.S3Client
+import spock.lang.Specification
+
+@Property(name = "aws.region", value = "eu-west-1")
+@MicronautTest(startApplication = false)
+class Apache5ServiceClientSpec extends Specification {
+    @Inject
+    BeanContext beanContext
+
+    void "a sync service client is built when the apache 5 client is the only sync client on the classpath"() {
+        when:
+        S3Client client = beanContext.getBean(S3Client)
+
+        then:
+        client.serviceName() == S3Client.SERVICE_NAME
+    }
+}
