+++ test-suite-groovy-aws-sdk-v2-http-client-netty/build.gradle.kts
@@ -1,0 +1,23 @@
+plugins {
+    id("groovy")
+    id("java-library")
+    id("io.micronaut.build.internal.aws-tests")
+}
+
+dependencies {
+    testCompileOnly(mn.micronaut.inject.groovy)
+    testImplementation(mnTest.micronaut.test.spock)
+    testImplementation(projects.micronautAwsSdkV2)
+    testImplementation(libs.awssdk.netty.nio.client)
+}
+
+tasks {
+    named("test", Test::class) {
+        useJUnitPlatform()
+    }
+}
+
+java {
+    sourceCompatibility = JavaVersion.toVersion("25")
+    targetCompatibility = JavaVersion.toVersion("25")
+}
