+++ test-suite-groovy-aws-sdk-v2-http-client-netty/src/test/groovy/io/micronaut/aws/sdk/v2/client/netty/NettyClientFactorySpec.groovy
