+++ test-suite-groovy-aws-sdk-v2-http-client-urlconnection/src/test/groovy/io/micronaut/aws/sdk/v2/client/urlconnection/UrlConnectionClientFactorySpec.groovy
